feat: extract useBox hook for better composability
