+++ lib/components/Box/Box.tsx
@@ -1,29 +1,6 @@
-import clsx from 'clsx';
-import React, { type ElementType, type ComponentPropsWithRef } from 'react';
+import React, { type ElementType } from 'react';
 
-import { useDeepCompareMemo } from '../../hooks';
-import type {
-	Sprinkles,
-	SprinklesResponsive,
-	SprinklesLegacyColours,
-} from '../../styles/sprinkles.css';
-import { filterNonSprinklesProps } from '../../utils/sprinkles';
-
-import { boxStyles } from './boxStyles';
-
-export type StyleProps = Sprinkles &
-	SprinklesResponsive &
-	SprinklesLegacyColours;
-type PolymorphicComponentProps<E extends ElementType, Props = object> = {
-	as?: E;
-	className?: clsx.ClassValue;
-} & Omit<ComponentPropsWithRef<E>, keyof Props | 'as' | 'className'> &
-	Props;
-
-export type BoxProps<E extends ElementType = 'div'> = PolymorphicComponentProps<
-	E,
-	StyleProps
->;
+import { useBox, type UseBoxProps } from './useBox';
 
 /**
  * A polymorphic Box component that provides a flexible container with styling capabilities, defauilting to a `<div>` element.
@@ -43,31 +20,10 @@ export type BoxProps<E extends ElementType = 'div'> = PolymorphicComponentProps<
  * @example
  * <Box display={['block', 'flex']} p={['3', '6', '8']}>Responsive padding</Box>
  */
-export const Box = <E extends ElementType = 'div'>({
-	as,
-	className,
-	...props
-}: BoxProps<E>) => {
-	const Component = as ?? 'div';
+export const Box = <E extends ElementType = 'div'>(props: UseBoxProps<E>) => {
+	const { Component, componentProps } = useBox<E>(props);
 
-	// Deep compare props for style calculation
-	const style = useDeepCompareMemo(
-		() =>
-			boxStyles({
-				as,
-				className,
-				...props,
-			}),
-		[as, className, props],
-	);
-
-	// Deep compare props for remaining props
-	const remainingProps = useDeepCompareMemo(
-		() => filterNonSprinklesProps(props),
-		[props],
-	);
-
-	return <Component className={style} {...remainingProps} />;
+	return <Component {...componentProps} />;
 };
 
 Box.displayName = 'Box';
