+++ lib/components/Box/index.ts
@@ -1,2 +1,3 @@
 export * from './Box';
+export * from './useBox';
 export { boxStyles } from './boxStyles';
